books/requirements.txt
altgraph==0.17
future==0.18.2
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
pefile==2021.5.13
protobuf==3.17.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pywin32-ctypes==0.2.0
qt-material==2.8.8
six==1.16.0
